refactor: block valdiatoin and active block validatoins extracted to util files
